reorg with better namespacing
